remove redundant folders
