+++ lumigator/frontend/src/App.vue
@@ -61,8 +61,8 @@
 <script lang="ts" setup>
 import { onMounted, ref } from 'vue'
 import LMenu from '@/components/organisms/LMenu.vue'
-import { useDatasetStore } from '@/stores/datasets/datasetsStore'
-import { useExperimentStore } from '@/stores/experiments/experimentsStore'
+import { useDatasetStore } from '@/stores/datasetsStore'
+import { useExperimentStore } from '@/stores/experimentsStore'
 import { useSlidePanel } from '@/composables/SlidingPanel'
 import ConfirmDialog from 'primevue/confirmdialog'
 import Toast from 'primevue/toast'
