+++ lumigator/frontend/src/components/molecules/LExperimentDetails.vue
@@ -139,7 +139,7 @@
 <script lang="ts" setup>
 import { computed, ref } from 'vue'
 import { storeToRefs } from 'pinia'
-import { useExperimentStore } from '@/stores/experiments/experimentsStore'
+import { useExperimentStore } from '@/stores/experimentsStore'
 import { formatDate, calculateDuration } from '@/helpers/index'
 import Button from 'primevue/button'
 import Tag from 'primevue/tag'
