+++ lumigator/frontend/src/components/molecules/LInferenceJobsTable.vue
@@ -68,8 +68,8 @@ import Tag from 'primevue/tag'
 import Column from 'primevue/column'
 import { formatDate } from '@/helpers/index'
 import { storeToRefs } from 'pinia'
-import { useExperimentStore } from '@/stores/experiments/experimentsStore'
-import { useDatasetStore } from '@/stores/datasets/datasetsStore'
+import { useExperimentStore } from '@/stores/experimentsStore'
+import { useDatasetStore } from '@/stores/datasetsStore'
 import { useSlidePanel } from '@/composables/SlidingPanel'
 
 const experimentStore = useExperimentStore()
