+++ lumigator/frontend/src/components/organisms/LDatasetDetails.vue
@@ -108,12 +108,12 @@
 <script lang="ts" setup>
 import { ref } from 'vue'
 import { storeToRefs } from 'pinia'
-import { useDatasetStore } from '@/stores/datasets/datasetsStore'
+import { useDatasetStore } from '@/stores/datasetsStore'
 import { useSlidePanel } from '@/composables/SlidingPanel'
 import { formatDate } from '@/helpers/index'
 import Button from 'primevue/button'
 import LGenerateGroundTruthPopup from '@/components/molecules/LGenerateGroundTruthPopup.vue'
-import { useExperimentStore } from '@/stores/experiments/experimentsStore'
+import { useExperimentStore } from '@/stores/experimentsStore'
 
 const datasetStore = useDatasetStore()
