+++ lumigator/frontend/src/main.ts
@@ -7,7 +7,7 @@ import { router } from '@/router'
 import PrimeVue from 'primevue/config'
 import ConfirmationService from 'primevue/confirmationservice'
 import ToastService from 'primevue/toastservice'
-import { LumiPreset } from './libs/primevue.config'
+import { LumiPreset } from './primevue.config'
 
 const app = createApp(App)
